+++ podchat/src/main/java/com/fanap/podchat/chat/ChatListener.java
@@ -372,4 +372,5 @@ public interface ChatListener {
     default void OnTagParticipantAdded(String content, ChatResponse<TagParticipantResult> response){}
     default void OnTagParticipantRemoved(String content, ChatResponse<TagParticipantResult> response){}
     default void OnTagList(String content, ChatResponse<TagListResult> response){}
+    default void onGetMutualGroups(String content, ChatResponse<ResultThreads> thread){}
 }