+++ podchat/src/main/java/com/fanap/podchat/requestobject/RequestGetContact.java
@@ -4,17 +4,31 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
 public class RequestGetContact extends BaseRequestObject {
+    String username;
 
     RequestGetContact(@NonNull Builder builder) {
         super(builder);
+        if (builder.username != null)
+            this.username = builder.username;
     }
 
-    public static class Builder extends BaseRequestObject.Builder<Builder>{
+    public String getUsername() {
+        return username;
+    }
 
-         @NonNull
-         public RequestGetContact build(){
-           return new RequestGetContact(this);
-         }
+
+    public static class Builder extends BaseRequestObject.Builder<Builder> {
+        String username = null;
+
+        @NonNull
+        public RequestGetContact build() {
+            return new RequestGetContact(this);
+        }
+
+        public Builder setUserName(String username) {
+            this.username = username;
+            return this;
+        }
 
         @Nullable
         @Override